+++ php/WpDateTimeZone.php
@@ -10,14 +10,14 @@ class WpDateTimeZone extends \DateTimeZone {
 	/**
 	 * Determine time zone from WordPress options and return as object.
 	 *
-	 * @return self
+	 * @return static
 	 */
 	public static function getWpTimezone() {
 
 		$timezone_string = get_option( 'timezone_string' );
 
 		if ( ! empty( $timezone_string ) ) {
-			return new self( $timezone_string );
+			return new static( $timezone_string );
 		}
 
 		$offset  = get_option( 'gmt_offset' );
@@ -25,6 +25,6 @@ class WpDateTimeZone extends \DateTimeZone {
 		$minutes = ( $offset - floor( $offset ) ) * 60;
 		$offset  = sprintf( '%+03d:%02d', $hours, $minutes );
 
-		return new self( $offset );
+		return new static( $offset );
 	}
 }