+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## Fixed
+- Instance `static` (not `self`) for correct behavior on subclassing.
+
 ## 0.1 - 2016-11-15
 
 ### Added